+++ src/DatasetBase/Contract/CollectionInterface.php
@@ -54,4 +54,17 @@ interface CollectionInterface extends \IteratorAggregate, \Countable, \ArrayAcce
     public function map(callable $mapFn): iterable;
 
     public function column(string $valueAccessor, ?string $keyAccessor = null): iterable;
+
+    /**
+     * Create a new collection of the same type, but without any content.
+     */
+    public function withoutItems(): self;
+
+    /**
+     * Group items in maximum $size big chunks. The last chunk can be less than $size items.
+     *
+     * @psalm-param positive-int $size
+     * @psalm-return iterable<self&non-empty-list<T>>
+     */
+    public function chunk(int $size): iterable;
 }