+++ test/DatasetBase/CollectionTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Heptacom\HeptaConnect\Dataset\Base\Test;
 
 use Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract;
+use Heptacom\HeptaConnect\Dataset\Base\ScalarCollection\IntegerCollection;
 use Heptacom\HeptaConnect\Dataset\Base\Test\Fixture\SerializationDatasetEntity;
 use Heptacom\HeptaConnect\Dataset\Base\Test\Fixture\UsageStructCollection;
 use PHPUnit\Framework\TestCase;
@@ -12,6 +13,7 @@ use PHPUnit\Framework\TestCase;
 /**
  * @covers \Heptacom\HeptaConnect\Dataset\Base\Contract\DatasetEntityContract
  * @covers \Heptacom\HeptaConnect\Dataset\Base\DatasetEntityCollection
+ * @covers \Heptacom\HeptaConnect\Dataset\Base\ScalarCollection\IntegerCollection
  * @covers \Heptacom\HeptaConnect\Dataset\Base\Support\AbstractCollection
  * @covers \Heptacom\HeptaConnect\Dataset\Base\Support\AbstractObjectCollection
  * @covers \Heptacom\HeptaConnect\Dataset\Base\Support\JsonSerializeObjectVarsTrait
@@ -189,6 +191,26 @@ final class CollectionTest extends TestCase
             ],
         ]);
         static::assertCount(3, $collection);
+    }
+
+    public function testChunking(): void
+    {
+        $items = \range(1, 96);
+        $collection = new IntegerCollection($items);
+        $chunks = \iterable_to_array($collection->chunk(30));
+        $chunks = \array_map('iterable_to_array', $chunks);
+        static::assertCount(30, $chunks[0]);
+        static::assertCount(30, $chunks[1]);
+        static::assertCount(30, $chunks[2]);
+        static::assertCount(6, $chunks[3]);
+        static::assertSame(\array_chunk($items, 30), $chunks);
+
+        static::assertSame([], \iterable_to_array((new IntegerCollection())->chunk(10)));
+
+        $chunks = \iterable_to_array((new IntegerCollection([1, 2, 3]))->chunk(-5));
+        $chunks = \array_map('iterable_to_array', $chunks);
+
+        static::assertSame([[1], [2], [3]], $chunks);
     }
 
     /**