+++ src/DatasetBase/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add method `\Heptacom\HeptaConnect\Dataset\Base\Contract\CollectionInterface::withoutItems` and implementation in `\Heptacom\HeptaConnect\Dataset\Base\Support\AbstractCollection::withoutItems` to create safely new instances of the same type but without content
+- Add method `\Heptacom\HeptaConnect\Dataset\Base\Contract\CollectionInterface::chunk` and implementation in `\Heptacom\HeptaConnect\Dataset\Base\Support\AbstractCollection::chunk` to iterate over the items prepared in a buffer of a certain size
+
 ### Changed
 
 ### Deprecated